+++ src/App.tsx
@@ -13,8 +13,7 @@ import SlidingPage from "@/ui/SlidingPage";
 import { useModalStore } from "@/lib/store/modalStore";
 import { createPortal } from "react-dom";
 import { retrieveLaunchParams } from "@telegram-apps/sdk-react";
-import { telegramActionGetDialogs } from "@/lib/telegram/api/telegramActions";
-import { TelegramDialogInfo } from "@/lib/telegram/api/telegramApiClient";
+import { getTelegramDialogsAndSetToStore } from "./helpers/telegram/getTelegramDialogsAndSetToStore";
 
 export default function App() {
   const { jazzProfile } = useJazzProfileContext();
@@ -125,35 +124,6 @@ const ModalsAndSlidingPages = () => {
       </SlidingPage>
     </div>
   );
-};
-
-const getTelegramDialogsAndSetToStore = async () => {
-  const { setTelegramDialogs } = useAppStore.getState();
-  const tempDialogs: TelegramDialogInfo[] = [];
-
-  await telegramActionGetDialogs().then((dialogs) => {
-    dialogs.forEach((d) => {
-      if (d.entity?.className === "User") {
-        if (!d.entity.username) return;
-        const userInfo = {
-          label: d.entity.firstName || "X",
-          username: d.entity.username || "X",
-          id: d.entity.id,
-        };
-        tempDialogs.push(userInfo);
-      } else if (d.entity?.className === "Channel") {
-        if (!d.entity.username) return;
-        const userInfo = {
-          label: d.entity.title || "X",
-          username: d.entity.username || "X",
-          id: d.entity.id,
-        };
-        tempDialogs.push(userInfo);
-      }
-    });
-  });
-
-  setTelegramDialogs(tempDialogs);
 };
 
 const setupTelegramTheme = (): void => {
